Move around more stuff!

+++ mifia/salem/salem.py
@@ -1,20 +1,18 @@
 from ..base import Game
-from .deaths import LynchedByTheTown
-import enum
+from ..base.errors import InvalidStateError
+from .deaths import LynchedByTheTown, LeftTheGame
+from .moment import Moment, GamePhase
+from .judgement import Judgement
 import typing
 if typing.TYPE_CHECKING:
     from multiprocessing import Queue
-    from ..base import Player, RoleList, NameList
-
-
-class Judgement(enum.IntEnum):
-    GUILTY = -1
-    ABSTAINED = 0
-    INNOCENT = 1
+    from ..base import RoleList, NameList
+    from .player import SalemPlayer
 
 
 class Salem(Game):
-    """A game mode where discussion takes place during the dawn, a player is voted during the day and a judgement is passed on them during the dusk.
+    """A game mode where discussion takes place during the dawn, a player is voted during the day and a judgement
+     is passed on them during the dusk.
      If they are found guilty, they are lynched, while if they are found innocent, nothing happens.
      In either case, the game moves on to the night, where nothing* happens.
 
@@ -22,12 +20,13 @@ class Salem(Game):
 
     def __init__(self, rolelist: "RoleList", namelist: "NameList", outgoing_queue: "Queue"):
         super().__init__(rolelist, namelist, outgoing_queue)
-        self.votes: typing.Dict[Player, Player] = {}
-        self.on_trial: typing.Optional[Player] = None
-        self.judgements: typing.Dict[Player, Judgement] = {}
+        self.votes: typing.Dict[SalemPlayer, SalemPlayer] = {}
+        self.on_trial: typing.Optional[SalemPlayer] = None
+        self.judgements: typing.Dict[SalemPlayer, Judgement] = {}
+        self.moment: Moment = None
 
-    def vote_count(self) -> typing.Dict[Player, int]:
-        counts: typing.Dict[Player, int] = {}
+    def vote_count(self) -> typing.Dict[SalemPlayer, int]:
+        counts: typing.Dict[SalemPlayer, int] = {}
         for player in self.players.by_randomness():
             counts[player] = 0
         for player in self.votes:
@@ -44,12 +43,44 @@ class Salem(Game):
             score += self.judgements[player].value
         return score
 
-    def _end_day(self):
-        super()._end_day()
+    def end_dawn(self):
+        if self.moment is None:
+            raise InvalidStateError("Game is not GameState.IN_PROGRESS!")
+        if self.moment.phase != GamePhase.DAWN:
+            raise InvalidStateError("Game is not GamePhase.DAWN!")
+        self.moment = Moment(GamePhase.DAY, self.moment.cycle)
+        for player in self.players.by_priority():
+            player.role.on_day()
+            if not player.connected:
+                player.kill(LeftTheGame(self.moment))
+
+    def end_day(self):
+        if self.moment is None:
+            raise InvalidStateError("Game is not GameState.IN_PROGRESS!")
+        if self.moment.phase != GamePhase.DAY:
+            raise InvalidStateError("Game is not GamePhase.DAY!")
+        self.moment = Moment(GamePhase.DUSK, self.moment.cycle)
+        for player in self.players.by_priority():
+            player.role.on_dusk()
         self.on_trial = self.vote_order()[0]
 
-    def _end_dusk(self):
-        super()._end_dusk()
+    def end_dusk(self):
+        if self.moment is None:
+            raise InvalidStateError("Game is not GameState.IN_PROGRESS!")
+        if self.moment.phase != GamePhase.DUSK:
+            raise InvalidStateError("Game is not GamePhase.DUSK!")
+        self.moment = Moment(GamePhase.NIGHT, self.moment.cycle)
+        for player in self.players.by_priority():
+            player.role.on_night()
         if self.judgements_count() < 0:
             self.on_trial.kill(LynchedByTheTown(self.moment, self.judgements))
         self.on_trial = None
+
+    def end_night(self):
+        if self.moment is None:
+            raise InvalidStateError("Game is not GameState.IN_PROGRESS!")
+        if self.moment.phase != GamePhase.NIGHT:
+            raise InvalidStateError("Game is not GamePhase.NIGHT!")
+        self.moment = Moment(GamePhase.DAWN, self.moment.cycle + 1)
+        for player in self.players.by_priority():
+            player.role.on_dawn()
